third_party/WebKit/LayoutTests/crypto/aes-ctr-generateKey-parseAlgorithm-failures.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests generateKey() with bad AES-CTR parameters.");
jsTestIsAsync = true;

extractable = true;
keyUsages = ['encrypt', 'decrypt'];

Promise.resolve(null).then(function() {
    return crypto.subtle.generateKey({ name: 'aes-ctr' }, extractable, keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({ name: 'aes-ctr', length: 70000 }, extractable, keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({ name: 'aes-ctr', length: -3 }, extractable, keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({ name: 'aes-ctr', length: -Infinity }, extractable, keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
}).then(finishJSTest, failAndFinishJSTest);

</script>

</body>
</html>
