third_party/WebKit/LayoutTests/crypto/aes-ctr-generateKey-parseAlgorithm-failures-expected.txt

Tests generateKey() with bad AES-CTR parameters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


error is: TypeError: AesKeyGenParams: length: Missing required property
error is: TypeError: AesKeyGenParams: length: Outside of numeric range
error is: TypeError: AesKeyGenParams: length: Outside of numeric range
error is: TypeError: AesKeyGenParams: length: Outside of numeric range
PASS successfullyParsed is true

TEST COMPLETE
